+++ src/main/java/org/cubalibre/familink/api/web/controller/RestProfilsController.java
@@ -1,0 +1,35 @@
+package org.cubalibre.familink.api.web.controller;
+
+import org.cubalibre.familink.api.entite.Profil;
+import org.cubalibre.familink.api.services.IProfilService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/profils")
+public class RestProfilsController {
+
+    private static final Logger LOG = LoggerFactory.getLogger(RestProfilsController.class.getName());
+
+    @Autowired
+    @Qualifier("profilService")
+    private IProfilService profilService;
+
+
+    // ********** CREATE PROFILS ********** //
+    @RequestMapping(path = "/", method = RequestMethod.GET, produces = "application/json")
+    @ResponseBody
+    public List<Profil> getProfils() {
+
+        return profilService.getProfils();
+    }
+
+}
